refactor(terminal): extract TerminalWorkerIngestController
- Move the flag-isolated live worker-parse ingest (issue #10960) —
  applyWorkerIngestPolicy, createWorkerIngest, the workerIngest map, and
  the onWorkerIngestEngaged subscription — into a new sibling controller
- TerminalInstanceService reaches diversion state via getIngest(id) from
  the onData handler and delegates destroy(id)/dispose() cleanup
