+++ src/services/terminal/TerminalInstanceService.ts
@@ -53,13 +53,11 @@ import type { PaintSurface } from "./paintFabric/PaintSurfaceRegistry";
 import { createRoundRobinPlacement } from "./paintFabric/placementPolicies";
 import {
   isPaintFabricEnabled,
-  isPaintFabricWorkerIngestEnabled,
   getPaintFabricSurfaceCount,
   paintFabricAuxSurfaceId,
   PRIMARY_SURFACE_ID,
 } from "./paintFabric/paintFabricConfig";
-import { LiveWorkerIngest } from "./workerParse/LiveWorkerIngest";
-import { createParseWorkerTransport } from "./workerParse/createParseWorkerTransport";
+import { TerminalWorkerIngestController } from "./TerminalWorkerIngestController";
 import { PERF_MARKS } from "@shared/perf/marks";
 import { markRendererPerformance } from "@/utils/performance";
 import { safeFireAndForget } from "@/utils/safeFireAndForget";
@@ -81,10 +79,6 @@ const WEBGL_RESTORE_DEBOUNCE_MS = 100;
 // multi-terminal hide. Authoritative release paths (tier demotion, agent
 // demotion, destroy, hibernation) cancel this timer and release immediately.
 const WEBGL_HIDE_DWELL_MS = 500;
-
-// Poll interval while the worker-ingest engage barrier waits for the normal
-// pipeline (ingest queue + pending xterm writes) to quiesce (#10960).
-const WORKER_INGEST_DRAIN_POLL_MS = 10;
 
 function canAutoInitializeTerminalIngest(): boolean {
   return (
@@ -129,12 +123,8 @@ class TerminalInstanceService {
   private settleWaiters: TerminalSettleWaiterRegistry;
   private burstController: TerminalBurstController;
   private resizePassScheduler: TerminalResizePassScheduler;
+  private workerIngestController: TerminalWorkerIngestController;
   private unsubTierChanged: (() => void) | null = null;
-  // Live worker-parse ingest (issue #10960), gated by
-  // DAINTREE_PAINT_FABRIC_WORKER_INGEST. One controller per terminal that has
-  // ever gone BACKGROUND while the gate is on; disposed with the terminal.
-  private workerIngest = new Map<string, LiveWorkerIngest>();
-  private unsubWorkerIngestEngaged: (() => void) | null = null;
 
   constructor() {
     if (canAutoInitializeTerminalIngest()) {
@@ -175,6 +165,15 @@ class TerminalInstanceService {
       getInstance: (id) => this.instances.get(id),
       isResizeLocked: (id) => this.resizeController.isResizeLocked(id),
       resize: (id, width, height) => this.resizeController.resize(id, width, height),
+    });
+
+    this.workerIngestController = new TerminalWorkerIngestController({
+      getInstance: (id) => this.instances.get(id),
+      getQueuedBytes: (id) => this.dataBuffer.getQueuedBytes(id),
+      resumeFlush: (id) => this.dataBuffer.resumeFlush(id),
+      incrementUnseen: (id, isScrolledBack) =>
+        this.unseenTracker.incrementUnseen(id, isScrolledBack),
+      fetchAndRestore: (id) => this.restoreController.fetchAndRestore(id),
     });
 
     this.writeController = new TerminalWriteController({
@@ -261,7 +260,7 @@ class TerminalInstanceService {
         // mode via the release path above) — push a fresh diagnostics sample.
         this.scheduleWhySlowReport();
 
-        this.applyWorkerIngestPolicy(id, tier, managed);
+        this.workerIngestController.applyWorkerIngestPolicy(id, tier, managed);
       },
     });
 
@@ -1546,7 +1545,7 @@ class TerminalInstanceService {
       // controller — it acks them immediately and lands them on the mirror
       // through the session's zero-loss buffering, never through the write
       // controller.
-      const ingest = this.workerIngest.get(id);
+      const ingest = this.workerIngestController.getIngest(id);
       if (ingest?.shouldDivert()) {
         ingest.feedDiverted(data);
         return;
@@ -3089,128 +3088,6 @@ class TerminalInstanceService {
     return this.webGLManager.isActive(id);
   }
 
-  /**
-   * Tier-driven worker-ingest policy (issue #10960): BACKGROUND demotes the
-   * terminal's parse into a worker via its dedicated pty-host port; any other
-   * tier (FOCUSED/BURST, and VISIBLE — a visible pane repainting at snapshot
-   * cadence would be a UX regression) promotes back to the main-thread path
-   * through the zero-loss release barrier. No-op unless
-   * DAINTREE_PAINT_FABRIC_WORKER_INGEST=1.
-   */
-  private applyWorkerIngestPolicy(
-    id: string,
-    tier: TerminalRefreshTier,
-    managed: ManagedTerminal
-  ): void {
-    if (!isPaintFabricWorkerIngestEnabled()) return;
-    if (tier === TerminalRefreshTier.BACKGROUND) {
-      let ingest = this.workerIngest.get(id);
-      if (!ingest) {
-        ingest = this.createWorkerIngest(id, managed);
-        this.workerIngest.set(id, ingest);
-      }
-      ingest.setDesired(true);
-    } else {
-      this.workerIngest.get(id)?.setDesired(false);
-    }
-  }
-
-  private createWorkerIngest(id: string, managed: ManagedTerminal): LiveWorkerIngest {
-    if (!this.unsubWorkerIngestEngaged) {
-      this.unsubWorkerIngestEngaged = terminalClient.onWorkerIngestEngaged((terminalId) => {
-        this.workerIngest.get(terminalId)?.handleEngaged();
-      });
-    }
-
-    const utf8ByteLength = (data: string): number => new TextEncoder().encode(data).byteLength;
-
-    const ingest = new LiveWorkerIngest({
-      id,
-      requestPort: () => terminalClient.requestWorkerIngestPort(id),
-      releasePort: () => terminalClient.releaseWorkerIngestPort(id),
-      sendEngage: () => terminalClient.sendWorkerIngestEngage(id),
-      sendRelease: (drainId) => terminalClient.sendWorkerIngestRelease(id, drainId),
-      createTransport: () => createParseWorkerTransport(),
-      mirror: {
-        write: (data, callback) => {
-          const current = this.instances.get(id);
-          if (!current) {
-            callback?.();
-            return;
-          }
-          // Direct write — snapshot applies and replays are pre-acked, so the
-          // write controller's ack bookkeeping must never see them. Unseen
-          // tracking still counts each repaint as output activity.
-          this.unseenTracker.incrementUnseen(id, current.isUserScrolledBack);
-          current.terminal.write(data, callback);
-        },
-      },
-      serializeMirror: () => {
-        const current = this.instances.get(id);
-        return current?.serializeAddon.serialize() ?? "";
-      },
-      getGeometry: () => {
-        const current = this.instances.get(id);
-        return {
-          cols: current?.terminal.cols ?? 80,
-          rows: current?.terminal.rows ?? 24,
-          scrollback: current?.terminal.options.scrollback ?? 1000,
-        };
-      },
-      ackDiverted: (data) => {
-        // Delivery source is encoded in the chunk type (the same invariant
-        // TerminalWriteController's ackBytes rule rests on): port chunks are
-        // always Uint8Array, IPC chunks always strings. Settle exactly one
-        // port-ack FIFO entry per port chunk; never let an IPC chunk shift a
-        // port entry — that would prematurely ack a port chunk still queued.
-        if (typeof data === "string") {
-          terminalClient.acknowledgeData(id, utf8ByteLength(data));
-        } else {
-          terminalClient.acknowledgePortData(id, data.byteLength, 1);
-        }
-      },
-      drainPendingWrites: async () => {
-        // Quiesce the normal pipeline before the engage serialize: queued
-        // ingest chunks, in-flight xterm writes, and serialized-restore
-        // deferrals all still land on the mirror through the write controller
-        // — the serialize must happen after them or they vanish at the next
-        // snapshot apply. Diversion is already on, so no new inflow.
-        const deadline = Date.now() + 5000;
-        for (;;) {
-          const current = this.instances.get(id);
-          if (!current) return false;
-          if (
-            this.dataBuffer.getQueuedBytes(id) === 0 &&
-            (current.pendingWrites ?? 0) === 0 &&
-            !current.isSerializedRestoreInProgress
-          ) {
-            return true;
-          }
-          if (Date.now() > deadline) return false;
-          this.dataBuffer.resumeFlush(id);
-          await new Promise((resolve) => setTimeout(resolve, WORKER_INGEST_DRAIN_POLL_MS));
-        }
-      },
-      requestHostRestore: () => {
-        safeFireAndForget(this.restoreController.fetchAndRestore(id), {
-          context: "worker-ingest fallback restore",
-        });
-      },
-      onDidFallback: (reason) => {
-        logWarn("Worker ingest fell back to main-thread parse", { id, reason });
-      },
-    });
-
-    // Geometry follows the mirror in every mode so demotion re-seeds at the
-    // right size (session forwards resizes only while in worker mode).
-    const resizeDisposable = managed.terminal.onResize(({ cols, rows }) =>
-      ingest.resize(cols, rows)
-    );
-    managed.listeners.push(() => resizeDisposable.dispose());
-
-    return ingest;
-  }
-
   destroy(id: string): void {
     // Cancel an in-flight creation for this id: if the panel is torn down while
     // getOrCreate is still awaiting setupTerminalAddons, the instance isn't in
@@ -3252,14 +3129,7 @@ class TerminalInstanceService {
     this.resizeController.clearResizeJob(managed);
     this.resizeController.clearResizeLock(id);
     this.resizeController.clearSettledTimer(id);
-    // Worker ingest dies with the terminal: dispose terminates the Worker and
-    // closes the dedicated port on every teardown path (kill/trash/project
-    // close/LRU eviction) — a leaked producer port queues unboundedly (#6283).
-    const ingest = this.workerIngest.get(id);
-    if (ingest) {
-      ingest.dispose();
-      this.workerIngest.delete(id);
-    }
+    this.workerIngestController.destroy(id);
     // Renderer-side destroy without a prior kill (project close, LRU
     // eviction of an exited terminal) must still drain the port-ack FIFO
     // before the held queue is wiped (#9910). kill/gracefulKill/trash clear
@@ -3347,8 +3217,7 @@ class TerminalInstanceService {
     this.stopPolling();
     this.unsubTierChanged?.();
     this.unsubTierChanged = null;
-    this.unsubWorkerIngestEngaged?.();
-    this.unsubWorkerIngestEngaged = null;
+    this.workerIngestController.dispose();
     this.resizePassScheduler.dispose();
     this.reflowController.dispose();
     this.reconciliationWatchdog.dispose();
